many to many issue fix

+++ src/main/java/com/perennialsys/repository/HistoryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HistoryRepository extends JpaRepository<History,Integer> {
+public interface HistoryRepository extends JpaRepository<History,Long> {
 }
